+++ src/pie_modules/models/common/__init__.py
@@ -1,0 +1,1 @@
+from .model import Model
